Merge in MOB/vbhxh from feat/VBHXHMOB-168 to master_02

* commit '70552f136691438d5b1c9569e6a525beab5e2cf0':
  VBHXHMOB-168

+++ lib/modules/declare_info_630b/data/model/declare_info_630b_request_data.dart
@@ -34,7 +34,7 @@ class DeclareInfo630bRequestData {
   final String? denNgay;
 
   /// Tổng số ngày
-  final int? tongSoNgay;
+  final double? tongSoNgay;
 
   /// Từ ngày đơn vị
   final String? tuNgayDonVi;
